api/requirements.txt
fastapi
pytesseract
langchain-openai
Pillow
opencv-python-headless
python-dotenv
uvicorn
langchain
python-multipart
